[REF] sustainability: Migrate Button view management system to carbon.common.mixin
