+++ sustainability/models/carbon_mixin.py
@@ -47,11 +47,11 @@ _logger = logging.getLogger(__name__)
 
 class CarbonMixin(models.AbstractModel):
     _name = "carbon.mixin"
+    _inherit = ["carbon.common.mixin"]
     _description = "A mixin used to add carbon values on any model"
     _carbon_types = ["in", "out"]
     _fallback_records = []
     _carbon_enable_page = True
-    _carbon_enable_button = True
 
     # TODO: Thinks about compute this from env['carbon.line.mixin']._get_computation_levels_mapping()
     @api.model
@@ -494,36 +494,7 @@ class CarbonMixin(models.AbstractModel):
                     for sheet in arch.xpath("//sheet"):
                         sheet.append(cls._carbon_generate_page_xml(without_page=True))
 
-            if cls._carbon_enable_button:
-                for button_box in arch.xpath("//div[@name='button_box']"):
-                    button_box.extend(cls._carbon_generate_button_xml())
-
         return arch, view
-
-    @api.model
-    def _carbon_get_button_list(cls) -> list[dict[str, Any]]:
-        """
-        Return a list of buttons to add to the view. The buttons are dict that contains:
-        - field: the field to display the value of the button. This shall contain the field name related to the qty field.
-        - icon: the icon to display on the button. Not required, default is "fa-leaf"
-        - action: the action to execute when the button is clicked. If not provided, this will be computed using the field name with .replace("_qty", "_ids") and prefixing it with "action_see_"
-        - string: the string to display on the button
-        You can also add as much key value pairs as you want, they will be added to the button as attributes.
-
-        We will check if the field exists on the model, and if it doesn't, we will not add the button to the list.
-        """
-        button_list = [
-            # Carbon origin button
-            dict(
-                field="carbon_line_origin_qty",
-                icon="fa-leaf",
-                # action="action_see_carbon_line_origin_ids",
-                string=_("Carbon Footprint"),
-                # invisible=False, # Always show the button
-            ),
-        ]
-
-        return button_list
 
     @api.model
     def _carbon_get_other_fields(cls):
@@ -537,61 +508,6 @@ class CarbonMixin(models.AbstractModel):
         If one of the group_name or group_string is not provided, the field will be displayed in the default group.
         """
         return []
-
-    @api.model
-    def _carbon_generate_button_xml(cls, model_name: str | None = None):
-        model_name = model_name or cls._name
-        if model_name not in cls.env:
-            raise UserError(_("Model %s not found", model_name))
-        model = cls.env[model_name]
-
-        button_list = model._carbon_get_button_list()
-        field_name_to_button = {}
-
-        button_required_fields = ["field", "string"]
-        for button_dict in button_list:
-            if any(field not in button_dict for field in button_required_fields):
-                raise UserError(_("Button %s is missing required fields", button_dict))
-
-            field = button_dict.pop("field")
-            if field not in model._fields:
-                continue
-            if field in field_name_to_button:
-                raise UserError(_("Field %s is used by multiple buttons", field))
-
-            button = etree.Element("button", name=f"sustainability_button_{field}")
-            button.set("icon", button_dict.pop("icon", "fa-leaf"))
-            button.set("type", "object")
-            button.set("invisible", f"{field} < 1")
-            button.set("class", f"oe_stat_button {button_dict.pop('class', '')}")
-
-            for key, value in button_dict.items():
-                button.set(key, str(value))
-
-            field_name_to_button[field] = button
-
-            action_method_name = f"action_see_{field.replace('_qty', '_ids')}"
-            if button_dict.get("action"):
-                button.set("name", button_dict.pop("action"))
-            elif hasattr(model, action_method_name):
-                button.set("name", action_method_name)
-            else:
-                _logger.warning(
-                    f"Action {action_method_name} not found on model {model_name}"
-                )
-                continue
-
-            div = etree.SubElement(
-                button, "div", **{"class": "o_field_widget o_stat_info"}
-            )
-            span = etree.SubElement(div, "span", **{"class": "o_stat_value"})
-            etree.SubElement(
-                span, "field", **{"name": field, "nolabel": "1", "widget": "statinfo"}
-            )
-            span = etree.SubElement(div, "span", **{"class": "o_stat_text"})
-            span.text = button_dict.pop("string")
-
-        return list(field_name_to_button.values())
 
     @api.model
     def _carbon_generate_page_xml(
